Compatability with RP&R
Added checks for RP&R and MeanerMonsters presence, which is then used to activate code for that.
Canged the method ArmorMaterialIdentifier to use a formula to arrive at the same numbers as the mod originally used if-statements for. This allows for any future changes to armor values or new armor items to be seamlessly used by this mod.