Messing With Idea of Footstep Sounds, 2/14/2024
Messing With Idea of Footstep Sounds. Didn't get much done today, but decided maybe to try messing with adding in custom footstep sounds that change based on armor worn and such, maybe even dripping water after leaving a body of water, that sort of thing. This is very similar to the feature in the "Better Ambience" mod by Joshcamas, but different enough I figure I'd work on my own implementation. Though, I can already tell my way of doing it is probably going to be more messy, but I'll see, and if it works I can always clean it up later if I feel necessary, this is mainly a warming up thing to get started on the mod at all, so will see how it goes, 2/14/2024.