Started Making Plate Footstep Sounds & More Testing, 2/22/2024
Started Making Plate Footstep Sounds & More Testing. Finally started on the more "finalized" versions of some of the footstep sounds, and sort of came to a conclusion of how I will likely implement all this in the end. End result for now should probably be 60 or 70 audio-files, if my plan goes correct, rather than 200+ or something silly, but will have to see, for now just working on the audio files and testing each one as I go, next will be more types for the armor of the boots and such, so yeah, 2/22/2024.