Getting Back To Implementing Stuff In Code, 3/5/2024
Getting Back To Implementing Stuff In Code. Took a few weeks mostly procrastinating honestly, but very slowly working on many of the footstep sounds and such. Now that I have many of those done, I'm starting to work more on the code aspect again, this is just a small start from that, 3/5/2024.