More Logic, Some Testing, And Bug Fixes, 3/9/2024
More Logic, Some Testing, And Bug Fixes. Did a good amount of testing today, and seems to be mostly working pretty well for the exteriors so far, atleast about as well as I expected. Next time I work on this I'm going to try and get it working similarly for interiors as well, atleast the "climate footsteps" part, the armor related stuff I'm going to work on more later after this is in a decent state, not bad progress though, 3/9/2024.