Changed From Switch-Cases To Look-up Tables For Climate Footsteps, 3/11/2024
Changed From Switch-Cases To Look-up Tables For Climate Footsteps. Interkarma mentioned "Look-up Tables" or LUT and I decided to try and use them for the first time, in this case for the climate based footsteps and the tile-sets for said climates and such, seems to work fine, and is much smaller in size vertically, which I like alot, so will keep these in mind for the future. Tomorrow going to try working on the interior "climate" footsteps and such, 3/11/2024.