Footsteps Now Accurate With Building Interior Floors, 3/14/2024
Footsteps Now Accurate With Building Interior Floors. Was an interest approach to get this working, but it seems to be working pretty fine all things considered. Next I'm not sure what to work on, possibly do the same for dungeons, or try and get walking on non-ground things to be detected as well for a footstep sound change, as well as for swimming and such, will see, 3/14/2024.