Created & Imported Armor Swaying Sounds, 3/29/2024
Created & Imported Armor Swaying Sounds. Just created the sounds for this "feature" today, but good progress honestly, fairly happy with the sounds themselves. Still need to be tested and implemented in code though, but I'll do that tomorrow most likely. I'm considering maybe determining which "sway" sounds to used based on the armor being worn on the body, all besides the feet-slot, then taking the ones that are either the "loudest" or just whichever has the most numerous, will see, 3/29/2024.