Just Quick Backup Commit For Experimenting, 3/30/2024
Just Quick Backup Commit For Experimenting. So I got this part of the sway logic technically working fine, but I'm not too happy with the result, so making this commit just in-case something gets messed up after doing a most likely refactoring of the logic to hopefully make it more to what I expect from this feature, 3/30/2024.