Started Reworking PCO Original Old Code, 4/17/2024
Started Reworking PCO Original Old Code. So this took forever to get started on, but I'm finally getting some work done on it atleast, and so far the progress was very good today, got alot of stuff done, much more than expected, will see how it continues to develop over the next few days, happy with the progress so far, 4/17/2024.