More Stuff With Blocking & Armor Done For The Time Being, 4/28/2024
More Stuff With Blocking & Armor Done For The Time Being. Still obviously needs alot of testing and refinement, but atleast going forward more, next maybe actually using the DT and DR values to do something with and determine if an attack is negated fully, by what, and if not how much damage is done, etc, 4/28/2024.