Add Bunch Of Base Variables For Audio Implementation, 10/3/2024
Add Bunch Of Base Variables For Audio Implementation. This will be mostly for just having a place to put the placeholder sounds when I make them and use them for testing. Next time I work on this probably get the actual logic in there for when and how to play the sounds, then after that actually cut the placeholder clips into individually named files and import them, etc, 10/3/2024.