Experimenting With Ways To Handle Monster Weapons, 10/29/2024
Experimenting With Ways To Handle Monster Weapons. Alot of this will likely be changed or removed, I just have to think of a better way to handle these "special monster weapons" and somehow cast them into be recognizable as regular DFU items for most purposes, will see if that is possible though, will see, 10/29/2024.