More Playing Around With How To Implement Combat Sounds, 12/23/2024
More Playing Around With How To Implement Combat Sounds. Still need to test if these changes work. What I should do next after confirming that is getting some logic in that would allow sound clips to play when an attack kills the target of said attack. Currently it just gives the warning, "Can't play soundclip removed gameobject" or something like that. So what I think I'll do is to confirm doing math against the target's current health and the damage the attack is about to do, and if that value goes to 0 or below, then play the soundclip instead from the target's gameobject, but the position of that gameobject as a oneshot or whatever, I had to do similar for the chests in my other mod I believe, so should not be too difficult to get that working, hopefully, 12/23/2024.