Added Test Texture And Command To Open GUI, 12/27/2024
Added Test Texture And Command To Open GUI. Decided to add a test texture for the new "PCO Info GUI" I was considering adding. Also made a simple command to open the window in-game, next time I'll try adding more functions to it, such as the windows showing currently equipped items in the respective boxes, similar to the vanilla inventory, will have to see how that goes and use the vanilla window as an example for shrinking down the item textures to fit the boxes and such, 12/27/2024.