Mostly Done On Basic Durability Bar Length & Color Logic, 12/30/2024
Mostly Done On Basic Durability Bar Length & Color Logic. Just decided to do a few if-else statements to determine the colors, rather than something more complex like a gradient system or whatever. Next time I'll actually implement the value for the bars reflecting the current condition of the item in that slot, as well as remove the bar if that slot is empty and such, 12/30/2024.