Think I'm Finished With The Item Durability Bars For Now, 1/2/2025
Think I'm Finished With The Item Durability Bars For Now. Fairly happy with how these have came out so far, next I'll probably try to work on the display text listing the attributes and such for each item slot, and maybe the "extra info" window doing similar when hovering over something, similar to that window in the vanilla inventory widow, 1/2/2025.