Got Slot Selection Borders To Toggle Correctly, 2/1/2025
Got Slot Selection Borders To Toggle Correctly. Now that I have that aspect done, next time I should try to get the panels set up for the "additional equipment panels" that will actually show up giving a list of the valid equipment for that slot within your inventory, and all the logic to have those enabled and hide when toggled and on the correct side of the screen, etc, 2/1/2025.