Got Right & Left Equip Slot Pop-ups To Display & Toggle, 2/4/2025
Got Right & Left Equip Slot Pop-ups To Display & Toggle. Now that I have them displaying in the correct positions and toggling on and off when I want them to, depending on the currently slot selected, if any. Next I should get the actual logic and stuff working for these pop-ups, so they display the items and are clickable and function and such, that would be primarily implementing that custom "PCOItemListScroller" class I have made and changing that for the purposes I need it for in this case, so will see about that next time I work on this, not bad progress though considering the glacial speed I've been working on this, 2/4/2025.