Got Equipping/Unequipping/Swapping Items Functioning Now, 2/20/2025
Got Equipping/Unequipping/Swapping Items Functioning Now. When you right click a slot it unequips whatever is occupying that slot, when you left click and item in the Item Scroller list it equips that item or swaps it with the current item occupying that slot, seems to work as expected after just a bit of testing, so that's cool. Next time I work on this, I'll probably try and see about adding mini-condition bars for items in the item scroller windows and see how that looks, maybe show the condition in the tool-tip text as well, will see, 2/20/2025.