Got Mini-Durability Bars Working In Item Scroller Panel, 2/21/2025
Got Mini-Durability Bars Working In Item Scroller Panel. Next time I work on this, I'll consider trying to fix that bug/oversight that allows the open when the console window is open and typing a command using the letters of the hotkey for that window, not certain how I might fix this, but will give it a try atleast. After that, see about working on an experimenting with another panel that will show a comparison between stats of the currently equipped item, and whatever item is being hovered over in the item scroller panel, similar to that stat comparisons function in say World of Warcraft, etc, 2/21/2025.