Very Good Progress On Item Comparison Panel Feature, 3/8/2025
Very Good Progress On Item Comparison Panel Feature. Tomorrow or next time I work on this. Make additional logic for when comparing an item that is going to cause both items currently in the hands to be unequipped. So like if you are wielding a sword and shield, but going to equip a staff or other 2-handed weapon, take the changes of both hands into consideration for the comparison gains and loses text. Same deal if you have an open left-hand and are looking at a 1-handed weapon, this would normally try to equip that in the open left-hand, rather than replace the item in the right-hand, so update the text based on that as well. Essentially make the stuff more context sensitive, I've gotten this far with this comparison text stuff, so this should go about as smoothly, hopefully. Thankfully this only concerns the right and left hand slots, rather than all slots, so will try and do that next time, so far so good on this feature, 3/8/2025.