Finally Some Headway On The Item Comparison Text Logic, 3/17/2025
Finally Some Headway On The Item Comparison Text Logic. Next time I work on this, after testing and such will hopefully be done with this item comparison stuff, for some reason it got me stuck for a bit here, mostly mental block on the logic part and procrastination, as usual, oh well, atleast some progress on it now, 3/17/2025.