Think I'm Finished With The Item Comparison Feature For Now, 3/21/2025
Think I'm Finished With The Item Comparison Feature For Now. I think for now atleast, this feature is good enough. I think I'll probably do a bit more testing with it, maybe do a little showcase video of this and the progress so far. Then after that likely start working on the final feature I had planned for this UI, that being the filter and sort buttons, will see how that goes though, so far so good atleast, took longer than expected though due to procrastination and such, 3/21/2025