Filtering Restricted Items In Equip GUI Now Working, 4/3/2025
Filtering Restricted Items In Equip GUI Now Working. Next time I work on this, I'll have to see how to actual "sort" the list of items, depending on the state of the sort buttons and such, might be strange, but hopefully won't be super difficult to figure out, will see, 4/3/2025.